+++ src/psysem/sem/inference/__init__.py
@@ -1,8 +1,7 @@
 from .basic import InferenceResult, ParameterInferenceEntry, estimate_parameter_inference
 
 __all__ = [
-    "InferenceResult",
-    "ParameterInferenceEntry",
-    "estimate_parameter_inference",
+	"InferenceResult",
+	"ParameterInferenceEntry",
+	"estimate_parameter_inference",
 ]
-