+++ src/psysem/sem/measurement/__init__.py
@@ -3,8 +3,8 @@ from .contracts import LoadingParameter, MeasurementDesign
 from .identification import check_measurement_identification
 
 __all__ = [
-    "LoadingParameter",
-    "MeasurementDesign",
-    "build_measurement_design",
-    "check_measurement_identification",
+	"LoadingParameter",
+	"MeasurementDesign",
+	"build_measurement_design",
+	"check_measurement_identification",
 ]